requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
Django==3.2.14
django-admin-honeypot==1.1.0
django-tempus-dominus==5.1.2.17
idna==3.3
Pillow==9.1.1
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
Unidecode==1.3.4
urllib3==1.26.9
